+++ packages/orq-rc/RELEASES.md
@@ -19,3 +19,13 @@ Based on:
 - [python v1.41.0-rc.15] packages/orq-rc
 ### Releases
 - [PyPI v1.41.0-rc.15] https://pypi.org/project/orq-poc-python-multi-env-version/1.41.0-rc.15 - packages/orq-rc
+
+## 2024-11-25 07:08:30
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.444.4 (2.461.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v1.42.0-rc.2] packages/orq-rc
+### Releases
+- [PyPI v1.42.0-rc.2] https://pypi.org/project/orq-poc-python-multi-env-version/1.42.0-rc.2 - packages/orq-rc